pyproject.toml
[project]
name = "telegram-sales-bot"
version = "1.0.0"
description = "Telegram Sales Agent for Bali Real Estate - AI-powered conversational bot"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "claude-agent-sdk>=0.1.19",
    "telethon>=1.42.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "pytz>=2025.2",
    "tiktoken>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
telegram-daemon = "telegram_sales_bot.core.daemon:main"
telegram-scheduler = "telegram_sales_bot.scheduling.polling_daemon:main"
telegram-registry = "telegram_sales_bot.registry.runner:main"
telegram-outreach = "telegram_sales_bot.registry.outreach:main"
telegram-db-init = "telegram_sales_bot.database.init:main"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_sales_bot"]
